# UIAutomation: Adjusted setUp() and setUpWithError() methods to speed up execution