--- a/SimplenoteUITests/EmailLogin.swift
+++ b/SimplenoteUITests/EmailLogin.swift
@@ -3,12 +3,12 @@ import XCTest
 class EmailLogin {
 
     class func open() {
-        app.buttons[uidButton_LogIn].tap()
-        app.buttons[uidButton_LogInWithEmail].tap()
+        app.buttons[UID.Button.LogIn].tap()
+        app.buttons[UID.Button.LogInWithEmail].tap()
     }
 
     class func close() {
-        let backButton = app.navigationBars[uidNavBar_LogIn].buttons[uidButton_Back]
+        let backButton = app.navigationBars[UID.NavBar.LogIn].buttons[UID.Button.Back]
         guard backButton.exists else { return }
 
         backButton.tap()
@@ -17,17 +17,17 @@ class EmailLogin {
     class func logIn(email: String, password: String) {
         enterEmail(enteredValue: email)
         enterPassword(enteredValue: password)
-        app.buttons[uidButton_LogIn].tap()
+        app.buttons[UID.Button.LogIn].tap()
     }
 
     class func enterEmail(enteredValue: String) {
-        let field = app.textFields[uidTextField_Email]
+        let field = app.textFields[UID.TextField.Email]
         field.tap()
         field.typeText(enteredValue)
     }
 
     class func enterPassword(enteredValue: String) {
-        let field = app.secureTextFields[uidTextField_Password]
+        let field = app.secureTextFields[UID.TextField.Password]
         field.tap()
         field.typeText(enteredValue)
     }