… SimplenoteUITests/SimplenoteUITests.swift
… SimplenoteUITests/Trash.swift