UI Tests execution sped up by 110 seconds (~466 seconds vs ~570 seconds)

### Fix
1. Adjusted `setUp()` and `setUpWithError()` methods to avoid app launch before every test
2. For this created method that attempts returning to 'base' screen for every test: `getToAllNotes()` (for Note Editor tests)

Also (not related to speed):
3. Created several methods like `AllNotesAssert.notesExist() `on top of `AllNotesAssert.noteExists()` to allow passing an array of note names instead of calling the method N times
4. Added guard statements to several old methods

### Test
1. Running all tests and seeing no new fails introduced, with execution time being within 460 sec range.

### Review
Only one developer and one designer are required to review these changes, but anyone can perform the review.

### Release
These changes do not require release notes.