@pachlava 2021-02-22
@mokagio thanks! I applied everything you have proposed. Actually every comment of yours made the code shorter, this I something I really like!